log-generator/requirements.txt
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.61.1
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
pillow==12.1.0
psutil==5.4.7
psycopg2-binary==2.9.11
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
stress==1.0.0
threadpoolctl==3.6.0
